todo_api_project/requirements.txt
# Project Requirements
python=3.6+
django=3.0
djangorestframework
